Palindrome
Palindrome is a program for my class that was a true challenge. I had to
(without using arrays or any functions she had not taught us)  to make a
program in java that would reverse a number and then determine whether
it is a palindrome or not. The lack of using arrays made this an actual
challenge. I also learned that the int variable type has some horrible
size constraints when using arbitrary numbers in this way. Comments will
come soon.